Updates from PR feedback
- Restore separate dose and pump event uploads
- Remove SyncPumpEvent
